Add DiagramModel class and refactor sld & nad models
